backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
openai==1.30.5
httpx==0.27.0
python-dotenv==1.0.1
django-filter==24.2
gunicorn==21.2.0
django-cors-headers==3.13.0